refactor: extract PERCEPTION_MEANINGFUL_DELTA constant and marketAverageRawVP helper
